Updated to python3; fixed string/character decoding errors
Because of changes to python, the decode_line routine was
running into errors. I'm not sure how this program ever worked,
actually, because I don't know that you could ever directly
add bytes and strings? However, now with the bytes decoded into
proper strings, all is well, and I get a clean listing from a
BASIC program I wrote and recorded in MAME to a .wav.